# portrait-studio/docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ComplianceResponse": {
        "additionalProperties": false,
        "properties": {
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "markdown",
          "path"
        ],
        "title": "ComplianceResponse",
        "type": "object"
      },
...
      "FreeResponse": {
        "additionalProperties": false,
        "properties": {
          "free_mb_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Mb After"
          },
          "free_mb_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Mb Before"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "free_mb_before",
          "free_mb_after"
        ],
        "title": "FreeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelsResponse": {
        "additionalProperties": false,
        "properties": {
          "folders": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Folders",
            "type": "object"
          }
        },
        "required": [
          "folders"
        ],
        "title": "ModelsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VramResponse": {
        "additionalProperties": false,
        "properties": {
          "comfy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comfy Version"
          },
          "free_mb": {
            "title": "Free Mb",
            "type": "integer"
          },
          "gpu": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu"
          },
          "torch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Torch"
          },
          "torch_free_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Torch Free Mb"
          },
          "torch_total_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Torch Total Mb"
          },
          "total_mb": {
            "title": "Total Mb",
            "type": "integer"
          },
          "used_mb": {
            "title": "Used Mb",
            "type": "integer"
          }
        },
        "required": [
          "gpu",
          "total_mb",
          "free_mb",
          "used_mb"
        ],
        "title": "VramResponse",
        "type": "object"
      },
      "VramTableEntry": {
        "additionalProperties": false,
        "properties": {
          "face_detailer": {
            "title": "Face Detailer",
            "type": "boolean"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "measured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured At"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "peak_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Mb"
          },
          "upscale": {
            "title": "Upscale",
            "type": "number"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "method",
          "width",
          "height",
          "upscale",
          "face_detailer",
          "peak_mb"
        ],
        "title": "VramTableEntry",
        "type": "object"
      },
      "VramTableResponse": {
        "additionalProperties": false,
        "properties": {
          "checkpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/VramTableEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "gpu": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu"
          },
          "lora": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lora"
          },
          "measured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured At"
          },
          "vram_total_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram Total Mb"
          }
        },
        "required": [
          "gpu",
          "vram_total_mb",
          "measured_at",
          "entries"
        ],
        "title": "VramTableResponse",
        "type": "object"
      },
      "WorkflowJsonResponse": {
        "additionalProperties": false,
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "nodes": {
            "additionalProperties": true,
            "title": "Nodes",
            "type": "object"
          },
          "titles": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Titles",
            "type": "object"
          }
        },
        "required": [
          "method",
          "file",
          "titles",
          "nodes"
        ],
        "title": "WorkflowJsonResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "同一の架空キャラクター（実写調・日本人・成人）で画像を生成・管理するローカル専用 API。認証: ヘッダ `X-API-Key`（`GET /api/health` を除く）。エラーは `{\"detail\": \"...\"}`（日本語）。",
    "title": "portrait-studio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/system/compliance": {
      "get": {
        "description": "docs/COMPLIANCE.md so the settings screen can show the legal / prohibited-use notes.",
        "operationId": "compliance_api_system_compliance_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compliance",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/free": {
      "post": {
        "operationId": "free_api_system_free_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Free",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/models": {
      "get": {
        "operationId": "models_api_system_models_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/vram": {
      "get": {
        "operationId": "vram_api_system_vram_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vram",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/vram-table": {
      "get": {
        "operationId": "vram_table_api_system_vram_table_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VramTableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vram Table",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/workflows/{method}": {
      "get": {
        "description": "Raw API-format workflow (for the 上級 panel: view / copy the JSON).",
        "operationId": "workflow_json_api_system_workflows__method__get",
        "parameters": [
          {
            "in": "path",
            "name": "method",
            "required": true,
            "schema": {
              "title": "Method",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowJsonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Json",
        "tags": [
          "system"
        ]
      }
    }
  }
}